ident: Add RSS initialism.
"RSS" is a valid initialism and should be treated as such. Add it to
a section that is separate from the common initialisms from golint,
to make it easier to keep those in sync with golint.

This will be used during generation of github.com/shurcooL/octiconssvg
package via dmitri.shuralyov.com/kebabcase package.