fix: Union type handling - duplication, pointers
This change attempts to address two problems with unmarshaling the
results of using GraphQL fragments into two different possible union
types. This will hereafter be referred to as "unmarshaling into a
union" to keep things short. Those two problems where:

1. If unmarshaling into a union, and the containing Go struct has fields
   which are pointers to the unioned structs, then the prior
   implementation would return an error stating `struct field for
   "<fieldname" doesn't exist in any of 3 places to unmarshal`
2. If unmarshaling into a union, all shared fields of the child union
   types would be populated in ALL the child fields, not just the child
   field which had a name maching `__typename`. This led to lots of
   duplicated data.

To accomplish this, I've tried to enrich the information being tracked
in the decoder "stacks" so that we can track the name of the field that
we're marshaling into and the `__typename` field value in the containing
struct.

I've added tests for both of the problems described above.

Based on discussion from here:
https://github.com/hasura/go-graphql-client/issues/175